--- log.js
@@ -1,5 +1,0 @@
-function Log(type, message) {
-  console.log(type + ": " + message)
-}
-
-export default Log
