+++ pages/DynamicOptions/Count/GratitudeCount/index.html
@@ -4,13 +4,13 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Quantidade de Gratidões</title>
-    <link rel="stylesheet" href="counts.css" />
-    <link rel="stylesheet" href="button.css" />
+    <link rel="stylesheet" href="../styles.css" />
+    <link rel="stylesheet" href="../../../Layout/Button/styles.css" />
   </head>
   <body>
     <div class="container">
       <section class="title">
-        <img src="./images/gratitude.png" alt="Imagem vetorial de um personagem meditando." />
+        <img src="../../../../images/gratitude.png" alt="Imagem vetorial de um personagem meditando." />
         <h2>Gratidão</h2>
       </section>
       <section class="content">
