+++ pages/DynamicOptions/Count/IntentionCount/index.html
@@ -4,13 +4,13 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Quantidade de Intenções</title>
-    <link rel="stylesheet" href="counts.css" />
-    <link rel="stylesheet" href="button.css" />
+    <link rel="stylesheet" href="../styles.css" />
+    <link rel="stylesheet" href="../../../Layout/Button/styles.css" />
   </head>
   <body>
     <div class="container">
       <section class="title">
-        <img src="./images/intention.png" alt="Imagem vetorial de uma caneta." />
+        <img src="../../../../images/intention.png" alt="Imagem vetorial de uma caneta." />
         <h2>Intenções<br>Diárias</h2>
       </section>
       <section class="content">
@@ -20,5 +20,5 @@
       </section>
     </div>
   </body>
-  <script src="count.js" type="module"></script>
+  <script src="../index.js" type="module"></script>
 </html>
