+++ pages/DynamicOptions/Identification/index.html
@@ -4,12 +4,12 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
     <title>Seu Nome Personalizado</title>
-    <link rel="stylesheet" href="identification.css" />
-    <link rel="stylesheet" href="button.css" />
+    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="../../Layout/Button/styles.css" />
   </head>
   <body>
     <section>
-      <img src="./images/userIcon.png" alt="Imagem vetorizada que representa um usuário qualquer." />
+      <img src="../../../images/userIcon.png" alt="Imagem vetorizada que representa um usuário qualquer." />
       <h1>Identificação</h1>
       <p>Para melhores experiências, queremos que digitemos seu nome (pode ser apelido, também!). Não se preocupe, o aplicativo não utilizará deste nome para nenhum dado sensível em nenhum momento.</p>
       <input type="text" name="name" id="name" placeholder="Seu nome vem aqui!" />
