+++ pages/DynamicOptions/Identification/styles.css
@@ -1,5 +1,5 @@
-@import 'initialConfig.css';
-@import 'fontsUrl.css';
+@import '../../../styles/initialConfig.css';
+@import '../../../styles/fontsUrl.css';
 
 body, section {
   display: flex;
