+++ pages/DynamicOptions/Pin/index.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-with, initial-scale=1.0" />
     <title>Seu PIN de Segurança</title>
-    <link rel="stylesheet" href="shared.css" />
-    <link rel="stylesheet" href="pin.css" />
-    <link rel="stylesheet" href="button.css" />
+    <link rel="stylesheet" href="../../../styles/shared.css" />
+    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="../../Layout/Button/styles.css" />
   </head>
   <body>
     <div class="top-svg">
@@ -14,7 +14,7 @@
     </div>
     <section>
       <section class="title">
-        <img src="./images/checked.png" alt="Imagem vetorial de uma marcação de correto." />
+        <img src="../../../images/checked.png" alt="Imagem vetorial de uma marcação de correto." />
         <h2>Código de<br>verificação</h2>
       </section>
       <section class="content">
@@ -36,5 +36,5 @@
       <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 320"><path fill="#820263" fill-opacity="1" d="M0,32L60,64C120,96,240,160,360,165.3C480,171,600,117,720,90.7C840,64,960,64,1080,69.3C1200,75,1320,85,1380,90.7L1440,96L1440,320L1380,320C1320,320,1200,320,1080,320C960,320,840,320,720,320C600,320,480,320,360,320C240,320,120,320,60,320L0,320Z"></path></svg>
     </div>
   </body>
-  <script src="pin.js" type="module"></script>
+  <script src="index.js" type="module"></script>
 </html>
