+++ pages/DynamicOptions/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8" />
     <title>Antes de prosseguirmos…</title>
     <link rel="stylesheet" href="styles.css" />
-    <link rel="stylesheet" href="button.css" />
+    <link rel="stylesheet" href="../Layout/Button/styles.css" />
   </head>
   <body>
     <div id="shape-tl">
