+++ pages/Intention/index.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width" />
     <title>Gratidões</title>
-    <link rel="stylesheet" href="dashboard.css" />
-    <link rel="stylesheet" href="header.css" />
-    <link rel="stylesheet" href="footer.css" />
+    <link rel="stylesheet" href="../Layout/Dashboard/styles.css" />
+    <link rel="stylesheet" href="../Layout/Header/styles.css" />
+    <link rel="stylesheet" href="../Layout/Footer/styles.css" />
   </head>
   <body>
     <header>
@@ -25,46 +25,46 @@
       </aside>
     </section>
      <footer>
-  <a href="dreamDiary.html">
+  <a href="../DreamDiary/index.html">
     <nav class="item">
       <img src="../../images/cloud.png" alt="Imagem vetorizada de uma nuvem" />
       <p>Diário de<br>Sonhos</p>
     </nav>
   </a>
-  <a href="routine.html">
+  <a href="../Routine/index.html">
     <nav class="item">
       <img src="../../images/clock.png" alt="Imagem vetorizada de um relógio" />
       <p>Rotina</p>
     </nav>
   </a>
-  <a href="dailyRegister.html">
+  <a href="../DailyRegister/index.html">
     <nav class="item">
       <img src="../../images/write.png" alt="Imagem vetorizada de uma mão com um lápis escrevendo" />
       <p>Registro diário</p>
     </nav>
   </a>
-  <a href="home.html">
+  <a href="../Home/index.html">
     <img src="../../images/lotus.png" alt="" />
   </a>
-  <a href="gratitude.html">
+  <a href="../Gratitude/index.html">
     <nav class="item">
       <img src="../../images/meditation.png" alt="Imagem vetorizada de um personagem meditando." />
       <p>Gratidão</p>
     </nav>
   </a>
-  <a href="intention.html">
+  <a href="../Intention/index.html">
     <nav class="item">
       <img src="../../images/dove.png" alt="Imagem vetorizada de um pombo" />
       <p>Intenção</p>
     </nav>
   </a>
-  <a href="distractions.html">
+  <a href="../Distractions/index.html">
     <nav class="item">
       <img src="../../images/distractions.png" alt="Imagem vetorizada de um cérebro com um raio na cabeça" />
       <p>Distrações</p>
     </nav>
   </a>
 </footer>
-<script src="dashboard.js" type="module"></script>
+<script src="../Layout/Dashboard/index.js" type="module"></script>
   </body>
 </html>
