+++ pages/Routine/index.html
@@ -5,8 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Configuração de Rotina</title>
     <link rel="stylesheet" href="routine.css" />
-    <link rel="stylesheet" href="footer.css" />
-    <link rel="stylesheet" href="header.css" />
+    <link rel="stylesheet" href="../Layout/Footer/styles.css" />
+    <link rel="stylesheet" href="../Layout/Header/styles.css" />
   </head>
   <body>
     <header>
@@ -27,7 +27,7 @@
     <main>
       <section class="morning">
         <section class="header">
-          <img src="./images/morning.png" alt="Image! vetorizada de um sol" />
+          <img src="../../images/morning.png" alt="Image! vetorizada de um sol" />
           <h2>Rotina da Manhã</h2>
         </section>
         <form>
@@ -40,7 +40,7 @@
       </section>
       <section class="afternoon">
         <section class="header">
-          <img src="./images/afternoon.png" alt="Imagem vetorizada de um sol entre nuvens" />
+          <img src="../../images/afternoon.png" alt="Imagem vetorizada de um sol entre nuvens" />
           <h2>Rotina da Tarde</h2>
         </section>
         <form>
@@ -53,7 +53,7 @@
       </section>
       <section class="night">
         <section class="header">
-          <img src="./images/night.png" alt="Image! vetorizada de um sol" />
+          <img src="../../images/night.png" alt="Image! vetorizada de um sol" />
           <h2>Rotina da Noite</h2>
         </section>
         <form>
@@ -67,46 +67,46 @@
       
     </main>
     <footer>
-  <a href="dreamDiary.html">
+  <a href="../DreamDiary/index.html">
     <nav class="item">
       <img src="../../images/cloud.png" alt="Imagem vetorizada de uma nuvem" />
       <p>Diário de<br>Sonhos</p>
     </nav>
   </a>
-  <a href="routine.html">
+  <a href="../Routine/index.html">
     <nav class="item">
       <img src="../../images/clock.png" alt="Imagem vetorizada de um relógio" />
       <p>Rotina</p>
     </nav>
   </a>
-  <a href="dailyRegister.html">
+  <a href="../DailyRegister/index.html">
     <nav class="item">
       <img src="../../images/write.png" alt="Imagem vetorizada de uma mão com um lápis escrevendo" />
       <p>Registro diário</p>
     </nav>
   </a>
-  <a href="home.html">
+  <a href="../Home/index.html">
     <img src="../../images/lotus.png" alt="" />
   </a>
-  <a href="gratitude.html">
+  <a href="../Gratitude/index.html">
     <nav class="item">
       <img src="../../images/meditation.png" alt="Imagem vetorizada de um personagem meditando." />
       <p>Gratidão</p>
     </nav>
   </a>
-  <a href="intention.html">
+  <a href="../Intention/index.html">
     <nav class="item">
       <img src="../../images/dove.png" alt="Imagem vetorizada de um pombo" />
       <p>Intenção</p>
     </nav>
   </a>
-  <a href="distractions.html">
+  <a href="../Distractions/index.html">
     <nav class="item">
       <img src="../../images/distractions.png" alt="Imagem vetorizada de um cérebro com um raio na cabeça" />
       <p>Distrações</p>
     </nav>
   </a>
 </footer>
-<script src="routine.js" type="module"></script>
+<script src="index.js" type="module"></script>
   </body>
 </html>
