+++ scripts/Dom/Set/index.js
@@ -1,0 +1,7 @@
+export function setAttributes(element, attributes) {
+    for (const key in attributes) {
+      element.setAttribute(key, attributes[key]);
+    }
+    return element
+  }
+
