+++ scripts/Encryption/index.js
